+++ changelog.md
@@ -1,5 +1,7 @@
 latest
  - Added Double selected action to add a copy of a sample to the end of itself for the selected samples.
+ - Prev / Next buttons on the edit panel to navigate between samples without needing to close the edit panel and re-select from the list.
+ - Cleaned up selected actions list, added overflow so list can scroll on mobile and flex to the width of the panel.
 
 2024.02.20 1.4.11
  - Fixes bug where saving slice data from imported slice file did not apply slice markers consistently.